Safari-611.1.10/LayoutTests/js/dom/modules/module-src-current-script.html

<!DOCTYPE HTML>
<html>
<head>
<script src="../../../resources/js-test-pre.js"></script>
</head>
<body>
<script>
description('Test document.currentScript should be null when executing the module with the src attribute.');

// Module will be executed asynchronously.
window.jsTestIsAsync = true;
</script>
<script>
debug('Module is not executed yet.');
</script>
<script src="../../../resources/js-test-post.js"></script>
<script type="module" src="./resources/module-src-current-script.js"></script>
</body>
</html>
