LayoutTests/js/dom/modules/module-src-current-script-expected.txt

Test document.currentScript should be null when executing the module with the src attribute.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Module is not executed yet.
Module execution is confined in the module environment.
PASS document.currentScript is null
PASS successfullyParsed is true

TEST COMPLETE
